+++ KoPrivateGPT/reranker/base.py
