+++ tests/KoPrivateGPT/llm/test_rerank_llm.py
@@ -5,7 +5,7 @@ import pytest
 
 import test_base_llm
 from KoPrivateGPT.llm.rerank import RerankLLM
-from KoPrivateGPT.utils.reranker import MonoT5Reranker
+from KoPrivateGPT.reranker import MonoT5Reranker
 
 logger = logging.getLogger(__name__)
 bm25_path = os.path.join(test_base_llm.root_dir, "resources", "bm25", "test_rerank_llm.pkl")
